+++ src/lib.rs
@@ -2,6 +2,7 @@
 
 pub mod btree;
 pub mod header;
+pub mod pager;
 pub mod record;
 pub mod schema;
 pub mod vfs;